Vukelic

+++ Self_Start_System/Client/src/app/forms/forms.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { FormService} from "../form.service";
 import { MatTableDataSource, MatSort } from '@angular/material';
 import {setUpFormContainer} from "@angular/forms/src/directives/shared";
+import { ConfirmDeleteDialogBoxComponent } from "../confirm-delete-dialog-box/confirm-delete-dialog-box.component";
+import { MatDialog, MatDialogRef, MAT_DIALOG_DATA} from "@angular/material";
 
 @Component({
   selector: 'app-forms',
@@ -15,7 +17,8 @@ export class FormsComponent implements OnInit {
   displayedColumns = ['formName', 'description', 'actionsColumn'];
   formDataSource;
 
-  constructor(private formService: FormService) { }
+  constructor(private formService: FormService,
+              public dialog: MatDialog) { }
 
   ngOnInit() {
     this.getAllForms();
@@ -43,13 +46,22 @@ export class FormsComponent implements OnInit {
   }
 
   deleteForm(form){
-    this.formService.deleteForm(form).subscribe(
-      res => {
-        console.log(res),
-          this.getAllForms()
-      },
-      error => console.log(error)
-    );
+
+    let dialogRef = this.dialog.open(ConfirmDeleteDialogBoxComponent, {
+      width: '250px',
+      data: {}
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      if(result){
+        this.formService.deleteForm(form).subscribe(
+          res => {
+            this.getAllForms();
+          },
+          error => console.log(error)
+        );
+      }
+    });
   }
 
   setUpDataSource = forms => {
